third_party/WebKit/LayoutTests/typedcssom/stylevalue-normalization/positionvalue-normalization-expected.txt

This is a testharness.js-based test.
FAIL CSS identifiers are left normalized to CSSKeywordValues assert_equals: expected "CSSPositionValue" but got "CSSStyleValue"
Harness: the test ran to completion.
